+++ src/Subscribers/AuthenticationLogSubscriber.php
@@ -7,25 +7,23 @@ use LangleyFoxall\LaravelAuthenticationLog\Models\AuthenticationLogRecord;
 
 class AuthenticationLogSubscriber
 {
-    // TODO pull from git repo to access createWithOmission method
+    // TODO pulled from master, watch out for merge conflict errors
     public function handleAuthenticatableLogin($event) 
     {
-        $fields = [
+        AuthenticationLogRecord::createWithConfigFilters([
             'authenticatable_id' => $event->user->id,
             'authenticatable_type' => get_class($event->user),
             'eventType' => get_class($event),
             'user_ip' => request()->getClientIp(),
-            'guard' => $event->guard,
+            'guard' => $event->guard, 
             'recorded_at' => now()
-        ];
-        
-        AuthenticationLogRecord::create(Omissions::omitFields($fields));
+        ]);
     }
 
     public function handleAuthenticatableFailed($event) 
     {
-        AuthenticationLogRecord::create([
-            'credentials' => Omissions::omitCredentials($event->credentials),
+        AuthenticationLogRecord::createWithConfigFilters([
+            'credentials' => $event->credentials,
             'eventType' => get_class($event),
             'user_ip' => request()->getClientIp(),
             'guard' => $event->guard,
@@ -35,7 +33,7 @@ class AuthenticationLogSubscriber
 
     public function handleAuthenticatableLogout($event) 
     {
-        AuthenticationLogRecord::create([
+        AuthenticationLogRecord::createWithConfigFilters([
             'authenticatable_id' => $event->user->id,
             'authenticatable_type' => get_class($event->user),
             'eventType' => get_class($event),
@@ -47,7 +45,7 @@ class AuthenticationLogSubscriber
 
     public function handleAuthenticatableRegistered($event) 
     {
-        AuthenticationLogRecord::create([
+        AuthenticationLogRecord::createWithConfigFilters([
             'authenticatable_id' => $event->user->id,
             'authenticatable_type' => get_class($event->user),
             'eventType' => get_class($event),
@@ -58,17 +56,17 @@ class AuthenticationLogSubscriber
 
     public function handleAuthenticatableLockout($event)
     {
-        AuthenticationLogRecord::create([
-            'credentials' => Omissions::omitCredentials($event->request->query()),
+        AuthenticationLogRecord::createWithConfigFilters([
+            'credentials' => $event->request->query(),
             'eventType' => get_class($event),
-            'user_ip' =>  $event->request->getClientIp,
+            'user_ip' =>  $event->request->getClientIp(),
             'recorded_at' => now(),
         ]);
     }
     
     public function handleAuthenticatablePasswordReset($event) 
     {   
-        AuthenticationLogRecord::create([
+        AuthenticationLogRecord::createWithConfigFilters([
                 'authenticatable_id' => $event->user->id,
                 'authenticatable_type' => get_class($event->user),
                 'eventType' => get_class($event),