+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 - Can now log ip addresses of all machines that triggered events
 - Can select fields to be omitted, for example, `user_ip`
-- Can not log User Registered events, Lockout events and Password Reset events
+- Can now log User Registered events, Lockout events and Password Reset events
+- Can now specify accepted guards within config file
+- Fixed field omissions not working correctly
+- Added `createWithConfigFilters` method to `AuthenticationLogRecord`
 
 ## v0.2